+++ src/services/index.ts
@@ -1,16 +1,5 @@
-export * from './getUsers';
-export * from './getUser';
-export * from './updateUser';
-export * from './getGroups';
-export * from './getGroup';
-export * from './updateGroup';
-export * from './createGroup';
-export * from './getApps';
-export * from './getApp';
-export * from './updateApplication';
-export * from './createApplication';
-export * from './getPolicy';
-export * from './getPolicies';
-export * from './createPolicy';
-export * from './updatePolicy';
-export * from './getApiKeys';
+export * from './UserService';
+export * from './GroupService';
+export * from './ApplicationService';
+export * from './PolicyService';
+export * from './ApiKeyService';
